refactor(core): extract notification tests
